feat: support in page searchbar
