+++ src/anotherWebBrowserIndex.ts
@@ -1,54 +1,14 @@
-import { clipboard } from 'electron';
-import {
-	App,
-	DropdownComponent,
-	Editor,
-	EventRef,
-	ItemView, MarkdownRenderer,
-	MarkdownView,
-	Menu,
-	Notice,
-	Plugin,
-	PluginSettingTab,
-	Setting,
-} from "obsidian";
-import { HeaderBar } from "./header_bar";
-import { FunctionHooks } from "./hooks";
-import { WEB_BROWSER_VIEW_ID, AnotherWebBrowserView } from "./web_browser_view";
-import { HTML_FILE_EXTENSIONS, WEB_BROWSER_FILE_VIEW_ID, AnotherWebBrowserFileView } from "./web_browser_file_view";
+import { App, Editor, EventRef, ItemView, MarkdownRenderer, MarkdownView, Menu, Notice, Plugin, } from "obsidian";
+import { HeaderBar } from "./component/headerBar";
+import { AnotherWebBrowserView, WEB_BROWSER_VIEW_ID } from "./web_browser_view";
+import { AnotherWebBrowserFileView, HTML_FILE_EXTENSIONS, WEB_BROWSER_FILE_VIEW_ID } from "./web_browser_file_view";
 import { t } from "./translations/helper";
 import { around } from "monkey-around";
 import { tokenType } from "./types/obsidian";
+import { AnotherWebBrowserPluginSettings, DEFAULT_SETTINGS, SEARCH_ENGINES, WebBrowserSettingTab } from "./setting";
+import { InPageSearchBar } from "./component/inPageSearchBar";
 
 // import { around } from "monkey-around";
-
-interface AnotherWebBrowserPluginSettings {
-	defaultSearchEngine: string;
-	customSearchUrl: string;
-	customHighlightFormat: boolean;
-	highlightFormat: string;
-	openInSameTab: boolean;
-	openInObsidianWeb: boolean;
-}
-
-const DEFAULT_SETTINGS: AnotherWebBrowserPluginSettings = {
-	defaultSearchEngine: 'duckduckgo',
-	customSearchUrl: 'https://duckduckgo.com/?q=',
-	customHighlightFormat: false,
-	highlightFormat: '[{CONTENT}]({URL})',
-	openInSameTab: false,
-	openInObsidianWeb: false,
-}
-
-// Add search engines here for the future used.
-export const SEARCH_ENGINES = {
-	'google': 'https://www.google.com/search?q=',
-	'bing': 'https://www.bing.com/search?q=',
-	'duckduckgo': 'https://duckduckgo.com/?q=',
-	'yahoo': 'https://search.yahoo.com/search?p=',
-	'baidu': 'https://www.baidu.com/s?wd=',
-	'wikipedia': 'https://en.wikipedia.org/w/index.php?search=',
-};
 
 export default class AnotherWebBrowserPlugin extends Plugin {
 	settings: AnotherWebBrowserPluginSettings;
@@ -69,15 +29,15 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 			new Notice(`File extensions ${ HTML_FILE_EXTENSIONS } had been registered by other plugin!`);
 		}
 
-		// FunctionHooks.onload();
 		this.updateEmptyLeaves(false);
 		this.registerContextMenu();
 		this.registerCustomURI();
 		this.dispatchMarkdownView();
+		this.dispatchWindowOpen();
 
 		this.onLayoutChangeEventRef = this.app.workspace.on("layout-change", () => {
 			const activeView = this.app.workspace.getActiveViewOfType(ItemView);
-			if (activeView) this.addHeader(activeView);
+			if (activeView) this.addHeaderAndSearchBar(activeView);
 		});
 
 		this.registerCommands();
@@ -85,7 +45,6 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 
 	onunload() {
 		this.app.workspace.detachLeavesOfType(WEB_BROWSER_VIEW_ID);
-		// FunctionHooks.onunload();
 		this.app.workspace.offref(this.onLayoutChangeEventRef);
 
 		// Clean up header bar added to "New tab" views when plugin is disabled.
@@ -94,7 +53,7 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 	}
 
 	// Add header bar to empty view.
-	private addHeader(currentView: ItemView) {
+	private addHeaderAndSearchBar(currentView: ItemView) {
 		if (!currentView) return;
 		// Check if new leaf's view is empty, else return.
 		if (currentView.getViewType() != "empty") return;
@@ -102,15 +61,25 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 		if (!currentView.headerEl.children[2].hasClass("web-browser-header-bar")) {
 			const headerBar = new HeaderBar(currentView.headerEl.children[2], this);
 			// Focus on current inputEl
-			headerBar.focus();
+			if (!this.settings.showSearchBarInPage) headerBar.focus();
 			headerBar.addOnSearchBarEnterListener((url: string) => {
+				AnotherWebBrowserView.spawnWebBrowserView(false, { url });
+			});
+		}
+		if (!currentView.contentEl.children[0].hasClass("wb-page-search-bar") && this.settings.showSearchBarInPage) {
+			const inPageSearchBar = new InPageSearchBar(currentView.contentEl.children[0], this);
+			if (currentView.contentEl.children[0].children[0]) {
+				(currentView.contentEl.children[0].children[0] as HTMLElement).style.opacity = "0.4";
+			}
+			inPageSearchBar.focus();
+			inPageSearchBar.addOnSearchBarEnterListener((url: string) => {
 				AnotherWebBrowserView.spawnWebBrowserView(false, { url });
 			});
 		}
 	}
 
 	// Clean up header bar added to empty views when plugin is disabled.
-	private removeHeader(currentView: ItemView) {
+	private removeHeaderAndSearchBar(currentView: ItemView) {
 		if (!currentView) return;
 		// Check if new leaf's view is empty, else return.
 		if (currentView.getViewType() != "empty") return;
@@ -119,6 +88,10 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 			currentView.headerEl.children[2].empty();
 			currentView.headerEl.children[2].removeClass("web-browser-header-bar");
 		}
+		if (currentView.contentEl.children[0].hasClass("wb-page-search-bar") && this.settings.showSearchBarInPage) {
+			currentView.contentEl.children[0].children[1].detach();
+			currentView.contentEl.children[0].removeClass("wb-page-search-bar");
+		}
 	}
 
 	// Update all leaf contains empty view when restart Obsidian
@@ -126,8 +99,8 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 		const emptyLeaves = this.app.workspace.getLeavesOfType("empty");
 		emptyLeaves.forEach((leaf) => {
 			if (leaf.view instanceof ItemView) {
-				if (!removeHeader) this.addHeader(leaf.view);
-				if (removeHeader) this.removeHeader(leaf.view);
+				if (!removeHeader) this.addHeaderAndSearchBar(leaf.view);
+				if (removeHeader) this.removeHeaderAndSearchBar(leaf.view);
 			}
 		});
 	}
@@ -194,7 +167,7 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 					// If checking is true, we're simply "checking" if the command can be run.
 					// If checking is false, then we want to actually perform the operation.
 					if (!checking) {
-						FunctionHooks.ogWindow$Open.call(window, webbrowserView.getState()?.url, "_blank");
+						window.open(webbrowserView.getState()?.url, "_blank");
 					}
 
 					// This command will only show up in Command Palette when the check function returns true
@@ -341,7 +314,9 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 					},
 			})
 		);
+	}
 
+	dispatchWindowOpen() {
 		// Use monkey-around to match current need.
 		// @ts-ignore
 		const uninstaller = around(window, {
@@ -356,11 +331,9 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 
 					if (decodeURI(urlString) !== urlString) urlString = decodeURI(urlString).toString().replace(/\s/g, "%20");
 
-					// 1. Allows Obsidian to open a popup window if url is "about:blank" and features is not null
-					// TODO: There might be a better way to detect if it's a popup window.
 					// 2. Perform default behavior if the url isn't "http://" or "https://"
 					// TODO: Change to `isWebUri()` when I change to use the valid-url library.
-					if ((urlString === "about:blank" && features) || (!urlString.startsWith("http://") && !urlString.startsWith("https://") && !(urlString.startsWith("file://") && /\.htm(l)?/g.test(urlString)))) {
+					if ((urlString === "about:blank" && features) || (!urlString.startsWith("http://") && !urlString.startsWith("https://") && !(urlString.startsWith("file://") && /\.htm(l)?/g.test(urlString))) || (urlString !== "about:blank" && target === "_blank")) {
 						return next(url, target, features)
 					}
 
@@ -370,7 +343,6 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 				}
 		})
 		this.register(uninstaller);
-
 	}
 
 	async loadSettings() {
@@ -379,157 +351,5 @@ export default class AnotherWebBrowserPlugin extends Plugin {
 
 	async saveSettings() {
 		await this.saveData(this.settings);
-	}
-}
-
-
-class WebBrowserSettingTab extends PluginSettingTab {
-	plugin: AnotherWebBrowserPlugin;
-	private applyDebounceTimer = 0;
-
-	constructor(app: App, plugin: AnotherWebBrowserPlugin) {
-		super(app, plugin);
-		this.plugin = plugin;
-	}
-
-	applySettingsUpdate() {
-		clearTimeout(this.applyDebounceTimer);
-		const plugin = this.plugin;
-		this.applyDebounceTimer = window.setTimeout(() => {
-			plugin.saveSettings();
-		}, 100);
-	}
-
-	display(): void {
-		const { containerEl } = this;
-
-		containerEl.empty();
-
-		containerEl.createEl('h2', { text: 'Web Browser' });
-
-		this.addSearchEngine();
-		this.addHighlightFormat();
-		this.addOpenInSameTab();
-		this.addOpenInObsidianWeb();
-	}
-
-	addSearchEngine() {
-		new Setting(this.containerEl)
-			.setName(t('Default Search Engine'))
-			.setDesc("")
-			.addDropdown(async (drowdown: DropdownComponent) => {
-				drowdown
-					.addOption('duckduckgo', t('DuckDuckGo'))
-					.addOption('google', t('Google'))
-					.addOption('bing', t('Bing'))
-					.addOption('yahoo', t('Yahoo'))
-					.addOption('baidu', t('Baidu'))
-					.addOption('custom', t('Custom'))
-					.setValue(this.plugin.settings.defaultSearchEngine).onChange(async (value) => {
-					this.plugin.settings.defaultSearchEngine = value;
-					this.applySettingsUpdate();
-					// Force refresh
-					this.display();
-				});
-			});
-
-		if (!(this.plugin.settings.defaultSearchEngine === 'custom')) {
-			return;
-		}
-
-		new Setting(this.containerEl)
-			.setName(t('Set Custom Search Engine Url'))
-			.setDesc(t("Set custom search engine url for yourself. 'Duckduckgo' By default"))
-			.addText((text) =>
-				text
-					.setPlaceholder(DEFAULT_SETTINGS.customSearchUrl)
-					.setValue(this.plugin.settings.customSearchUrl)
-					.onChange(async (value) => {
-						this.plugin.settings.customSearchUrl = value;
-						this.applySettingsUpdate();
-					}),
-			);
-	}
-
-	addHighlightFormat() {
-		new Setting(this.containerEl)
-			.setName(t('Custom Link to Highlight Format'))
-			.addToggle((toggle) => {
-				toggle
-					.setValue(this.plugin.settings.customHighlightFormat)
-					.onChange(async (value) => {
-						this.plugin.settings.customHighlightFormat = value;
-						this.applySettingsUpdate();
-						// Force refresh
-						this.display();
-					});
-			})
-
-		if (!this.plugin.settings.customHighlightFormat) {
-			return;
-		}
-
-		new Setting(this.containerEl)
-			.setName(t('Copy Link to Highlight Format'))
-			.setDesc(t("Set copy link to text fragment format. [{CONTENT}]({URL}) By default. You can also set {TIME:YYYY-MM-DD HH:mm:ss} to get the current date."))
-			.addText((text) =>
-				text
-					.setPlaceholder(DEFAULT_SETTINGS.highlightFormat)
-					.setValue(this.plugin.settings.highlightFormat)
-					.onChange(async (value) => {
-						if (value === "") {
-							this.plugin.settings.highlightFormat = DEFAULT_SETTINGS.highlightFormat;
-							this.applySettingsUpdate();
-							// Force refresh
-							this.display();
-						}
-						this.plugin.settings.highlightFormat = value;
-						this.applySettingsUpdate();
-					}),
-			);
-	}
-
-	addOpenInSameTab() {
-		new Setting(this.containerEl)
-			.setName(t('Open URL In Same Tab'))
-			.addToggle((toggle) => {
-				toggle.setValue(this.plugin.settings.openInSameTab)
-					.onChange(async (value) => {
-						this.plugin.settings.openInSameTab = value
-						this.applySettingsUpdate()
-						this.display()
-					})
-			})
-	}
-
-	addOpenInObsidianWeb() {
-		new Setting(this.containerEl)
-			.setName(t('Open URL In Obsidian Web'))
-			.addToggle((toggle) => {
-				toggle.setValue(this.plugin.settings.openInObsidianWeb)
-					.onChange(async (value) => {
-						this.plugin.settings.openInObsidianWeb = value;
-						this.applySettingsUpdate();
-						this.display();
-					})
-			})
-
-		if (!this.plugin.settings.openInObsidianWeb) {
-			return;
-		}
-
-
-		const bookmarkLetsEl = createEl("a", {
-			text: `Obsidian BookmarkLets Code`,
-			cls: 'cm-url',
-			href: 'javascript:(function(){var%20i%20%3Ddocument.location.href%3B%20document.location.href%3D%22obsidian%3A%2F%2Fweb-open%3Furl%3D%22%20%2B%20encodeURIComponent%28i%29%3B})();'
-		})
-		bookmarkLetsEl.addEventListener("click", () => {
-			clipboard.writeText(`javascript:(function(){var%20i%20%3Ddocument.location.href%3B%20document.location.href%3D%22obsidian%3A%2F%2Fweb-open%3Furl%3D%22%20%2B%20encodeURIComponent%28i%29%3B})();`)
-			new Notice(t("Copy BookmarkLets Success"))
-		})
-		this.containerEl.appendChild(bookmarkLetsEl
-		);
-
 	}
 }
