+++ styles.css
@@ -55,3 +55,24 @@
 	background-color: var(--background-modifier-form-field);
 	margin-left: 4px;
 }
+
+.wb-page-search-bar-input {
+	width: 500px;
+	min-width: 20px;
+	height: 44px !important;
+	border-radius: 15px !important;
+	margin-bottom: 20px;
+	margin-left: auto;
+	margin-right: auto;
+}
+
+.wb-page-search-bar {
+	flex-direction: column-reverse;
+}
+
+.wb-page-search-bar-text {
+	text-align: center;
+	margin-bottom: 20px;
+	font-size: 33px;
+	color: var(--color-purple);
+}
